:truck: Rename module, fix Logger type
