+++ tests/test_log_config.py
@@ -7,8 +7,8 @@ from unittest.mock import patch
 
 from loguru import logger
 
-from ohlc_toolkit.config import log_config
-from ohlc_toolkit.config.log_config import _get_log_file_path, get_logger
+from ohlc_toolkit.config import logging
+from ohlc_toolkit.config.logging import _get_log_file_path, get_logger
 
 
 class TestLogConfig(unittest.TestCase):
@@ -32,8 +32,8 @@ class TestLogConfig(unittest.TestCase):
     def test_serialization_format(self, mock_stdout):
         """Test that logs are serialized correctly."""
         with patch.dict(os.environ, {"ENABLE_SERIALIZE_LOGS": "TRUE"}):
-            importlib.reload(log_config)  # Reload the module to apply the env var
-            test_logger = log_config.get_logger("ohlc_toolkit.test_module2")
+            importlib.reload(logging)  # Reload the module to apply the env var
+            test_logger = logging.get_logger("ohlc_toolkit.test_module2")
             test_logger.bind(body="extra info").info("Test message")
 
             # Check that the serialized log is written to stdout
@@ -46,8 +46,8 @@ class TestLogConfig(unittest.TestCase):
     def test_serialization_format_with_exception(self, mock_stdout):
         """Test that logs with exceptions are serialized correctly."""
         with patch.dict(os.environ, {"ENABLE_SERIALIZE_LOGS": "TRUE"}):
-            importlib.reload(log_config)  # Reload the module to apply the env var
-            test_logger = log_config.get_logger("ohlc_toolkit.test_module2")
+            importlib.reload(logging)  # Reload the module to apply the env var
+            test_logger = logging.get_logger("ohlc_toolkit.test_module2")
 
             try:
                 raise ValueError("An error occurred")
@@ -70,15 +70,15 @@ class TestLogConfig(unittest.TestCase):
     def test_enable_file_logging(self, mock_add, _):
         """Test that file logging is enabled and configured correctly."""
         with patch.dict(os.environ, {"ENABLE_FILE_LOGGING": "TRUE"}):
-            importlib.reload(log_config)  # Reload the module to apply the env var
-            test_logger = log_config.get_logger("ohlc_toolkit.test_module")
+            importlib.reload(logging)  # Reload the module to apply the env var
+            test_logger = logging.get_logger("ohlc_toolkit.test_module")
             test_logger.info("Test message")
             mock_add.assert_any_call(
                 "/fake/path/to/logs/ohlc_toolkit/{time:YYYY-MM-DD}.log",
                 rotation="00:00",
                 retention="7 days",
                 enqueue=True,
-                level=log_config.FILE_LOG_LEVEL,
+                level=logging.FILE_LOG_LEVEL,
                 diagnose=True,
                 format=unittest.mock.ANY,
                 serialize=unittest.mock.ANY,
@@ -86,12 +86,12 @@ class TestLogConfig(unittest.TestCase):
 
     def side_effect_for_add(*args, **_):
         """Raise PermissionError on the second call."""
-        if hasattr(log_config, "call_count"):
-            log_config.call_count += 1
+        if hasattr(logging, "call_count"):
+            logging.call_count += 1
         else:
-            log_config.call_count = 1
+            logging.call_count = 1  # type: ignore
 
-        if log_config.call_count == 2:  # noqa: PLR2004
+        if logging.call_count == 2:  # type: ignore # noqa: PLR2004
             raise PermissionError("Mocked permission error")
 
     @patch("loguru._logger.Logger.add", side_effect=side_effect_for_add)
@@ -99,8 +99,8 @@ class TestLogConfig(unittest.TestCase):
     def test_file_logging_permission_error(self, mock_warning, _):
         """Test that a warning is logged when PermissionError occurs."""
         with patch.dict(os.environ, {"ENABLE_FILE_LOGGING": "TRUE"}):
-            importlib.reload(log_config)  # Reload the module to apply the env var
-            test_logger = log_config.get_logger("ohlc_toolkit.test_module")
+            importlib.reload(logging)  # Reload the module to apply the env var
+            test_logger = logging.get_logger("ohlc_toolkit.test_module")
 
             # Trigger a log to ensure the logger is initialized
             test_logger.info("Trigger log")
